refactor(daemon): co-locate all tests in src/ and merge 3 duplicate test files
Merge test/ versions (canonical, larger) with unique src/ tests into
single co-located files. Delete packages/daemon/test/ entirely.

- continuity-state.test.ts: 17 tests (11 from test/ + 6 from src/)
- session-checkpoints.test.ts: 28 tests (17 from test/ + 11 from src/)
- hooks.test.ts: 3502 lines (test/ base + 7 describe blocks + agent-scope writeMemoryMd)

Test count unchanged: 1959 tests across 175 files (was 178 with duplicates).
